refactor: 프로덕션/테스트 코드 패키지 분리
