+++ src/main/java/step1/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package step1;
+package step1.calculator;
 
 public class Calculator {
